+++ MoneyBuddi/src/dao/IBudgetDAO.java
@@ -14,7 +14,7 @@ public interface IBudgetDAO {
 	
 	void deleteBudget(Budget budget) throws SQLException;
 	
-	Budget getBudgetById(int id)throws Exception;
+	Budget getBudgetById(long id)throws Exception;
 	
 	Collection <Budget> getAllBudgetsForUser(User user) throws Exception;
